Move attribute id and its get/set methods to BaseBean class
