+++ Kitodo-DataManagement/src/main/java/org/kitodo/data/database/beans/Docket.java
@@ -15,8 +15,6 @@ import java.io.Serializable;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
 import javax.persistence.Table;
 
 @Entity
@@ -25,24 +23,11 @@ public class Docket extends BaseBean implements Serializable {
 
     private static final long serialVersionUID = -5187947220333984868L;
 
-    @Id
-    @Column(name = "id")
-    @GeneratedValue
-    private Integer id;
-
     @Column(name = "name")
     private String name;
 
     @Column(name = "file")
     private String file;
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
 
     public String getName() {
         return name;
